+++ multiboot2/src/builder/mod.rs
@@ -1,8 +1,8 @@
 //! Module for the builder-feature.
 
 mod information;
-pub(crate) mod traits;
 
+pub(crate) use information::AsBytes;
 pub use information::InformationBuilder;
 
 use alloc::alloc::alloc;
@@ -35,11 +35,10 @@ impl<T: TagTrait<Metadata = usize> + ?Sized> BoxedDst<T> {
     /// Create a boxed tag with the given content.
     ///
     /// # Parameters
-    /// - `typ` - The given [`TagTypeId`]
     /// - `content` - All payload bytes of the DST tag without the tag type or
     ///               the size. The memory is only read and can be discarded
     ///               afterwards.
-    pub(crate) fn new(typ: impl Into<TagTypeId>, content: &[u8]) -> Self {
+    pub(crate) fn new(content: &[u8]) -> Self {
         // Currently, I do not find a nice way of making this dynamic so that
         // also miri is guaranteed to be happy. But it seems that 4 is fine
         // here. I do have control over allocation and deallocation.
@@ -63,7 +62,7 @@ impl<T: TagTrait<Metadata = usize> + ?Sized> BoxedDst<T> {
         unsafe {
             // write tag type
             let ptrx = ptr.cast::<TagTypeId>();
-            ptrx.write(typ.into());
+            ptrx.write(T::ID.into());
 
             // write tag size
             let ptrx = ptrx.add(1).cast::<u32>();
@@ -110,6 +109,7 @@ impl<T: ?Sized + PartialEq> PartialEq for BoxedDst<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::TagType;
 
     const METADATA_SIZE: usize = 8;
 
@@ -128,6 +128,8 @@ mod tests {
     }
 
     impl TagTrait for CustomTag {
+        const ID: TagType = TagType::Custom(0x1337);
+
         fn dst_size(base_tag: &Tag) -> usize {
             assert!(base_tag.size as usize >= METADATA_SIZE);
             base_tag.size as usize - METADATA_SIZE
@@ -136,11 +138,10 @@ mod tests {
 
     #[test]
     fn test_boxed_dst_tag() {
-        let tag_type_id = 1337_u32;
         let content = "hallo";
 
-        let tag = BoxedDst::<CustomTag>::new(tag_type_id, content.as_bytes());
-        assert_eq!(tag.typ, tag_type_id);
+        let tag = BoxedDst::<CustomTag>::new(content.as_bytes());
+        assert_eq!(tag.typ, CustomTag::ID);
         assert_eq!(tag.size as usize, METADATA_SIZE + content.len());
         assert_eq!(tag.string(), Ok(content));
     }